Remove unportable short-casting hack
Comper compresses 16-bit values instead of 8-bit, unlike every other
format that this framework supports. Previously, this was supported
by casting the file buffer to an unsigned short pointer, but this is
unportable as it relies on short being 16-bit, which is not
guaranteed.

Instead, clownlzss now must be told how many bytes long a 'value' in
the uncompressed data is.