Speed-up compression considerably
clownlzss previously compressed in a very naive way, by scanning the
entire sliding window for matches. Now, however, it uses a hash table
combined with a series of linked lists so that all matches are
already known beforehand.

Haruhiko Okumura's 1989 LZSS compressor is very clever, and uses a
binary search tree to massively speed-up compression. Unfortunately,
a binary search tree is only useful for compressors that implement a
greedy compression algorithm, as its entire purpose is to skip
matches that aren't larger than the current largest one. This means
that clownlzss must use something else.

In particular, Okumura's compressor used a kind of hash table that
would point to a series of trees. By replacing the trees with linked
lists, this technique can be applied to clownlzss.

This technique allows clownlzss to track which strings start with
which byte. By doing this, clownlzss never has to scan for matches
for the current string: it will already have a list of every string
that is guaranteed to match it by at least one byte.