Absolutely horrific refactor.
The point of this is to eliminate the need for `memory_buffer.c`,
which implements a stretchy buffer. It's just lazy programming, so
I've replaced it with something a bit more C-ish... which required
rewriting everything.

I think I fixed a bug or two while I was at it, but I don't remember.

Now, the compressed data is streamed directly to the user, through
callbacks. I guess I could add support for streaming input too, but
that's for another time.